major refactor of 'check_understanding' component for sanity
